move all data variables to separate files
