+++ src/components/dashboard/Footer.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 
 export const Footer = () => (
-  <div className="">
+  <div>
     <footer className="flex-horizontal flex-align-items-center footer">
       <div className="copyright ">Copyright Abc 2015 &copy;</div>
       <ul className="flex-horizontal  footer-menu">
