+++ src/components/dashboard/Sales.tsx
@@ -1,92 +1,14 @@
 import * as React from "react";
 import ReactApexChart from "react-apexcharts";
-import { ApexOptions } from "apexcharts";
 
-const chartSeries: Array<any> = [
-  {
-    name: "TEAM A",
-    type: "line",
-    data: [30, 25, 36, 30, 45, 35, 64, 52, 59, 36, 39],
-  },
-];
+import { salesChartOptions, salesChartSeries } from "./data/salesChart";
 
-const chartOptions: ApexOptions = {
-  chart: {
-    height: 350,
-    type: "line",
-    stacked: false,
-    toolbar: {
-      show: false,
-    },
-    zoom: {
-      enabled: false,
-    },
-  },
-  stroke: {
-    width: [5],
-    curve: "smooth",
-  },
-  plotOptions: {
-    bar: {
-      columnWidth: "50%",
-    },
-  },
-  fill: {
-    opacity: [0.85, 0.25, 1],
-    gradient: {
-      inverseColors: false,
-      shade: "light",
-      type: "vertical",
-      opacityFrom: 0.85,
-      opacityTo: 0.55,
-      stops: [0, 100, 100, 100],
-    },
-  },
-  labels: [
-    "01/01/2003",
-    "02/01/2003",
-    "03/01/2003",
-    "04/01/2003",
-    "05/01/2003",
-    "06/01/2003",
-    "07/01/2003",
-    "08/01/2003",
-    "09/01/2003",
-    "10/01/2003",
-    "11/01/2003",
-  ],
-  markers: {
-    size: 0,
-  },
-  xaxis: {
-    type: "datetime",
-  },
-  yaxis: {
-    title: {
-      text: "Points",
-    },
-    min: 0,
-  },
-  tooltip: {
-    shared: true,
-    intersect: false,
-    y: {
-      formatter: function (yValue) {
-        if (typeof yValue !== "undefined") {
-          return yValue.toFixed(0) + " points";
-        }
-        return yValue;
-      },
-    },
-  },
-};
-
-const BarChart = () => (
+const SalesLineChart = () => (
   <div className="app">
     <div className="mixed-chart">
       <ReactApexChart
-        options={chartOptions}
-        series={chartSeries}
+        options={salesChartOptions}
+        series={salesChartSeries}
         type="line"
         height="280"
       />
@@ -108,7 +30,7 @@ export const Sales = () => (
     </div>
 
     <div className="second-column">
-      <BarChart />
+      <SalesLineChart />
     </div>
   </div>
 );
