+++ src/components/dashboard/data/followers.ts
@@ -1,0 +1,76 @@
+import { User } from "../../UserInformation";
+
+export const followers: User[] = [
+  {
+    id: 1,
+    status: "idle",
+    name: "Chris Fox",
+    role: "Designer, Blogger",
+  },
+  {
+    id: 2,
+    status: "idle",
+    name: "Mohmed Said",
+    role: "UI/UX Developer",
+  },
+  {
+    id: 3,
+    status: "active",
+    name: "Mogen Polish",
+    role: "Art Director, Movie cutter",
+  },
+  {
+    id: 4,
+    status: "idle",
+    name: "Chris Fox",
+    role: "Designer, Blogger",
+  },
+  {
+    id: 5,
+    status: "idle",
+    name: "Mohmed Said",
+    role: "UI/UX Developer",
+  },
+  {
+    id: 6,
+    status: "active",
+    name: "Mogen Polish",
+    role: "Art Director, Movie cutter",
+  },
+  {
+    id: 7,
+    status: "idle",
+    name: "Chris Fox",
+    role: "Designer, Blogger",
+  },
+  {
+    id: 8,
+    status: "idle",
+    name: "Mohmed Said",
+    role: "UI/UX Developer",
+  },
+  {
+    id: 9,
+    status: "active",
+    name: "Mogen Polish",
+    role: "Art Director, Movie cutter",
+  },
+  {
+    id: 10,
+    status: "idle",
+    name: "Chris Fox",
+    role: "Designer, Blogger",
+  },
+  {
+    id: 11,
+    status: "idle",
+    name: "Mohmed Said",
+    role: "UI/UX Developer",
+  },
+  {
+    id: 12,
+    status: "active",
+    name: "Mogen Polish",
+    role: "Art Director, Movie cutter",
+  },
+];
