+++ src/components/index.tsx
@@ -1,6 +1,7 @@
 import * as React from "react";
 import { useState } from "react";
 import clsx from "clsx";
+
 import { Menu } from "./dashboard/Menu";
 import { Navbar } from "./dashboard/Navbar";
 import { WelcomeToInfinity } from "./dashboard/WelcomeToInfinity";
